+++ CHANGELOG.md
@@ -4,7 +4,8 @@ PDBTools.jl Changelog
 Version 3.1.0
 -------------
 - ![FEATURE][badge-feature] Selection syntax now supports parenthesis and the shortcut for `or`, as in `resname ARG CYS ALA`. 
-- ![FIX][badge-fix] CYS is now classified as a polar residue, following the definition of VMD.
+- ![INFO][badge-info] CYS is now classified as a polar residue, following the definition of VMD.
+- ![INFO][badge-info] GLY is now classified as a polar residue, following the definition of VMD.
 
 Version 3.0.0
 -------------